requirements.txt
ultralytics>=8.0.0
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.18.5
opencv-python>=4.1.2
matplotlib>=3.2.2
pandas>=1.1.4
Pillow>=7.1.2
scikit-learn>=0.19.1
tqdm>=4.64.0
pyyaml>=5.3.1
tensorboard>=2.4.1
seaborn>=0.11.0
imgaug>=0.4.0 